chore(2.0): Search & Filter API changes
Flarum 2.0 introduces a search driver implementation and moves gambits to the frontend.
