+++ js/src/common/extend.ts
@@ -1,3 +1,4 @@
+import BestAnswerGambit from './gambits/BestAnswerGambit';
 import Extend from 'flarum/common/extenders';
 import Tag from 'ext:flarum/tags/common/models/Tag';
 
@@ -5,4 +6,5 @@ export default [
   new Extend.Model(Tag) //
     .attribute<boolean>('isQnA')
     .attribute<boolean>('reminders'),
+  new Extend.Search().gambit('discussions', BestAnswerGambit),
 ];
