+++ js/src/common/index.js
@@ -1,0 +1,5 @@
+import app from 'flarum/common/app';
+
+app.initializers.add('', () => {
+  console.log('[] Hello, forum and admin!');
+});
