+++ resources/locale/en.yml
@@ -112,6 +112,10 @@ fof-best-answer:
         Visit the discussion here: {discussion_url}, and select the most appropriate post as the best answer.
 
         Thanks!
+  lib:
+    gambits:
+      solved-discussions:
+        key: solved
 
 fof-user-directory:
   lib:
