+++ src/Search/BestAnswerPostFilter.php
@@ -11,8 +11,8 @@
 
 namespace FoF\BestAnswer\Search;
 
-use Flarum\Filter\FilterInterface;
-use Flarum\Filter\FilterState;
+use Flarum\Search\Filter\FilterInterface;
+use Flarum\Search\SearchState;
 use Flarum\Tags\Tag;
 use Flarum\User\User;
 use Illuminate\Database\Query\Builder;
@@ -25,12 +25,12 @@ class BestAnswerPostFilter implements FilterInterface
         return 'is:solution';
     }
 
-    public function filter(FilterState $filterState, string $filterValue, bool $negate)
+    public function filter(SearchState $state, array|string $value, bool $negate): void
     {
-        $this->constrain($filterState->getQuery(), $filterState->getActor(), $negate);
+        $this->constrain($state->getQuery(), $state->getActor(), $negate);
     }
 
-    protected function constrain(Builder $query, User $actor, bool $negate)
+    protected function constrain(\Illuminate\Database\Eloquent\Builder $query, User $actor, bool $negate): void
     {
         // Join the `discussions` table to access `best_answer_post_id`.
         $query->join('discussions', 'posts.discussion_id', '=', 'discussions.id')
